requirements.txt
azureml-core==1.51.0
azureml-defaults==1.51.0
azureml-mlflow==1.51.0
azureml-telemetry==1.51.0
azure-ai-ml==1.9.0
azure-identity==1.14.0
debugpy==1.6.3
fsspec==2022.1.0
ipykernel==6.0
matplotlib==3.7.2
numpy==1.21.5
pandas==2.0.3
pre-commit==2.15.0
psutil==5.9.5
PyYAML==6.0.1
scikit-learn==1.3.0
scipy==1.7.3
streamlit==1.27.0
tqdm==4.65.0
